+++ app/src/main/java/com/improve/latetrain/DrawerActivity.kt
@@ -1,8 +1,6 @@
 package com.improve.latetrain
 
 import android.os.Bundle
-import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.android.material.snackbar.Snackbar
 import androidx.core.view.GravityCompat
 import androidx.appcompat.app.ActionBarDrawerToggle
 import android.view.MenuItem
@@ -11,21 +9,36 @@ import com.google.android.material.navigation.NavigationView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import android.view.Menu
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.FirebaseDatabase
+import com.google.firebase.database.ValueEventListener
+import kotlinx.android.synthetic.main.content_drawer.*
 
 class DrawerActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
 
 
+    private val instance = FirebaseDatabase.getInstance()
+    private val totalMinutesLate = instance.getReference(FirebaseInfo.TOTAL_TIME_PATH)
+
+    private val onNavigationItemSelectedListener = BottomNavigationView.OnNavigationItemSelectedListener { item ->
+        val fragment = when (item.itemId) {
+            R.id.navigation_add_mins -> AddMinsFragment.newInstance()
+            R.id.navigation_history -> HistoryFragment.newInstance()
+            R.id.navigation_complaints -> ComplaintsFragment.newInstance()
+            else -> ComplaintsFragment.newInstance()
+        }
+        supportFragmentManager.beginTransaction().add(R.id.fragmentContainer, fragment).commit()
+        true
+    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_drawer)
         val toolbar: Toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
 
-        val fab: FloatingActionButton = findViewById(R.id.fab)
-        fab.setOnClickListener { view ->
-            Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                .setAction("Action", null).show()
-        }
+
         val drawerLayout: DrawerLayout = findViewById(R.id.drawer_layout)
         val navView: NavigationView = findViewById(R.id.nav_view)
         val toggle = ActionBarDrawerToggle(
@@ -35,6 +48,20 @@ class DrawerActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelec
         toggle.syncState()
 
         navView.setNavigationItemSelectedListener(this)
+
+        //Bottom navigation view setup
+        val bottomNavView: BottomNavigationView = findViewById(R.id.bottom_nav_view)
+        bottomNavView.setOnNavigationItemSelectedListener(onNavigationItemSelectedListener)
+        bottomNavView.selectedItemId = R.id.navigation_complaints
+
+        totalMinutesLate.addValueEventListener(object : ValueEventListener {
+            override fun onCancelled(p0: DatabaseError) {}
+
+            override fun onDataChange(p0: DataSnapshot) {
+                //live_minutes.text = String.format(getString(R.string.live_info_display_text), p0.value)
+                live_minutes.text = p0.value.toString()
+            }
+        })
     }
 
     override fun onBackPressed() {
