+++ app/src/main/res/drawable/round_shape.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#221f73"/>
-    <corners android:radius="100dp"/>
+    <solid android:color="#3880EC"/>
+    <corners android:radius="50dp"/>
+
 
 </shape>
