+++ app/src/main/res/layout/content_drawer.xml
@@ -8,5 +8,66 @@
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
         tools:showIn="@layout/app_bar_drawer"
         tools:context=".DrawerActivity">
+    <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/topLayout"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            android:layout_marginTop="16dp"
+            android:layout_marginLeft="16dp"
+            android:layout_marginRight="16dp"
+            android:background="@drawable/live_bar"
+    >
+        <TextView
+                android:id="@+id/textlivebar_tv"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toTopOf="@id/topLayout"
+                app:layout_constraintRight_toRightOf="@id/topLayout"
+                android:textColor="@android:color/white"
+                android:text="כמות הדקות שרכבת ישראל עיכבה אותנו"
+                android:fontFamily="@font/levenim_mt"
+                android:textSize="19sp"
+
+        />
+        <TextView
+                android:id="@+id/live_minutes"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toBottomOf="@+id/textlivebar_tv"
+                app:layout_constraintRight_toLeftOf="@+id/topLayout"
+                app:layout_constraintLeft_toRightOf="@+id/topLayout"
+                tools:text="5000"
+                android:textSize="30sp"
+                android:fontFamily="@font/levenim_mt"
+                android:textColor="#f9c871"
+        />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <FrameLayout
+            android:layout_width="0dp" android:layout_height="0dp"
+            android:id="@+id/fragmentContainer"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/bottom_nav_view"
+            app:layout_constraintTop_toBottomOf="@id/topLayout"
+    />
+
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+            android:id="@+id/bottom_nav_view"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="0dp"
+            android:layout_marginStart="0dp"
+            android:visibility="visible"
+            android:background="?android:attr/windowBackground"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:menu="@menu/bottom_nav_menu"/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
