+++ app/src/main/res/values-iw/strings.xml
@@ -5,4 +5,20 @@
     <string name="title_home">בית</string>
     <string name="title_notifications">עדכונים</string>
     <string name="live_info_display_text">"שידור חי: הרכבת עיכבה אותנו ב %1$d  דקות!</string>
+    <string name="your_location">מיקומך</string>
+    
+    <string name="title_activity_drawer">DrawerActivity</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">Android Studio</string>
+    <string name="nav_header_subtitle">android.studio@android.com</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="action_settings">Settings</string>
+
+    <string name="menu_home">Home</string>
+    <string name="menu_gallery">Gallery</string>
+    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_tools">Tools</string>
+    <string name="menu_share">Share</string>
+    <string name="menu_send">Send</string>
 </resources>
