+++ app/src/main/res/values-v21/styles.xml
@@ -1,5 +1,5 @@
 <resources>
-    <style name="NoActionBar" parent="Theme.Design.Light.NoActionBar">
+    <style name="AppTheme.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
