+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,7 @@
     <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
     <string name="live_info_display_text">Live: we have been delayed for %1$d minutes</string>
+    <string name="your_location">מיקומך</string>
     <string-array name="stations">
         <item>לאן פניך מועדות?</item>
         <item>אופקים</item>
